Move MessageList, MessageGroup and MessageIdOperations to operations module
